Projects/wallet_system/wallet_app/__init__.py
from flask import Flask
from flask_sqlalchemy import SQLAlchemy
from flask_jwt_extended import JWTManager

db = SQLAlchemy()
jwt = JWTManager()

def create_app(config_class='config.Config'):
    app = Flask(__name__)
    app.config.from_object(config_class)
    
    # Initialize extensions
    db.init_app(app)
    jwt.init_app(app)
    
    # Register blueprints
    from wallet_app.routes.transaction import transaction_bp
    from wallet_app.routes.auth import auth_bp
    from wallet_app.routes.user import user_bp
    
    app.register_blueprint(transaction_bp, url_prefix='/api/transaction')
    app.register_blueprint(auth_bp, url_prefix='/api/auth')
    app.register_blueprint(user_bp, url_prefix='/api/user')
    
    return app