Projects/wallet_system/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
python-dotenv==0.19.0
passlib==1.7.4
pyjwt==2.3.0